requirements.txt
numpy>=1.26.0
faiss-cpu>=1.8.0
boto3>=1.34.0
pytest>=8.0.0
sentence-transformers>=3.0.0
transformers>=4.40.0
python-dotenv>=1.0.1
